requirements.txt
numpy
scipy
pillow
psutil
PyQt6
PySide6_Fluent_Widgets
pywin32
trimesh
win32mica
PyOpenGL
PyOpenGL-accelerate
paramiko
opencv-python
natsort